requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.0
beautifulsoup4==4.12.2
requests==2.31.0

